requirements.txt
aws-cdk-lib==2.194.0
constructs>=10.0.0,<11.0.0
pytest==8.3.5
pytest-cov==3.0.0
syrupy==4.9.1
cfn-lint==1.35.1
black==25.1.0
pyyaml==6.0.1
pre-commit==3.7.1
